first constraints working
